+++ src/compas_timber/design/__init__.py
@@ -5,10 +5,18 @@ from .workflow import TopologyRule
 from .workflow import JointDefinition
 from .workflow import FeatureDefinition
 from .workflow import DebugInfomation
+from .workflow import ContainerDefinition
 from .workflow import guess_joint_topology_2beams
 from .workflow import set_default_joints
 
 from .wall_from_surface import SurfaceModel
+from .wall_populator import WallPopulator
+from .wall_populator import WallPopulatorConfigurationSet
+from .wall_populator import WallSelector
+from .wall_populator import AnyWallSelector
+from .wall_details import LConnectionDetailA
+from .wall_details import LConnectionDetailB
+from .wall_details import TConnectionDetailA
 
 __all__ = [
     "CategoryRule",
@@ -19,6 +27,14 @@ __all__ = [
     "FeatureDefinition",
     "DebugInfomation",
     "SurfaceModel",
+    "WallPopulator",
+    "WallPopulatorConfigurationSet",
+    "WallSelector",
+    "AnyWallSelector",
+    "LConnectionDetailA",
+    "LConnectionDetailB",
+    "TConnectionDetailA",
     "guess_joint_topology_2beams",
     "set_default_joints",
+    "ContainerDefinition",
 ]